+++ CHANGELOG.md
@@ -14,6 +14,7 @@ backwards compatibility.
 * Added granular wavetable bank. Only to pure data for now.
 * Added sampled oscillator chip wavetable bank. Only to pure data for now.
 * Added string box wavetable bank. Only to pure data for now.
+* Added flute wavetable bank. Only to pure data for now.
 * Lowered LPF on wavetables to make the lower end thicker.
 * Use floats for audio buffers instead of u16.
 * Replace pulse wavetable in eurorack with FM and osc. chip.